start to implement new moves concept

+++ engine/BasicTypes.h
@@ -188,39 +188,39 @@ __declspec (align(1)) class Index
 public:
 
   Index() : index_(-1) {}
-  Index(int8 i) : index_(i) {}
+  Index(int i) : index_(i) {}
   Index(int x, int y) { index_ = x | (y<<3); }
   Index(char x, char y) { index_ = (x-'a') | ((y-'1')<<3); }
 
-  operator int8 () const
+  operator int () const
   {
     return index_;
   }
 
-  operator int8 & ()
+  operator int & ()
   {
     return index_;
   }
 
-  int8 x() const
+  int x() const
   {
     THROW_IF(index_ < 0, "try to get x of invalid index");
     return index_&7;
   }
 
-  int8 y() const
+  int y() const
   {
     THROW_IF(index_ < 0, "try to get y of invalid index");
     return (index_>>3)&7;
   }
 
-  void set_x(int8 x)
+  void set_x(int x)
   {
     index_ &= ~7;
     index_ |= x & 7;
   }
 
-  void set_y(int8 y)
+  void set_y(int y)
   {
     index_ &= 7;
     index_ |= (y&7)<<3;
